Fix staticcheck
